Rename to users.test.js
